refactor : 관리자 권한이 필요없는 사용자용 랜더링 페이지 테스트 패키지 및 클래스 분리, 유틸 테스트 패키지 추가
