+++ src/test/java/roomescape/util/UtilTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.util;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
